+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to Q1Pulse will be documented in this file.
 
+## \[0.8.0] - 2022-12-19
+
+- Update to qblox_instruments v0.8
+- Added set_freq
+- Changed arguments set_ph, set_ph_delta
+- Added chirp to sequencer
+
 ## \[0.7.1] - 2022-12-08
 
 - Improved communication (configuration/upload) speed by postponing error checks.